Split multi-plate files
If we are sent a 3mf file with multiple plates, split it into
multiple 1-plate files.  There does not appear to be a way to
select a plate from the display on the printer, and while we
can indicate which plate to print via mqtt, there is no way
for the slicer to tell us which plate the user wants to print.

Therefore, the most useful thing to do is to split the file
so the user can select which "plate" (file) from the screen.

If we are told to print, we will still print the first plate.